Move moc.rest into the haas package
