--- moc/__init__.py
@@ -1,9 +1,0 @@
-"""The ``moc`` package contains the former contents of ``moc-rest``::
-
-    https://github.com/CCI-MOC/moc-rest
-
-That library was split off from the main HaaS repo because we had plans to
-use it in another codebase, but this never happened. It's been pulled back
-into HaaS, but thus far the directory structure has been left untouched. It
-will probably be reorganized in the near future.
-"""
