+++ tests/unit/rest.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from moc import rest
+from haas import rest
 
 from abc import ABCMeta, abstractmethod
 from StringIO import StringIO
